requirements.txt
eventlet==0.18.2
Flask==0.10.1
Flask-SocketIO==2.1
greenlet==0.4.9
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
paho-mqtt==1.1
psycopg2==2.6.1
python-engineio==0.8.7
python-socketio==1.0
pytz==2015.7
six==1.10.0
Werkzeug==0.11.3
wheel==0.24.0
